MDL-80710 stylelint Disable vendor specific rules.
The recommended configuration assumes the use of autoprefixer, which we
don't use. Those rules need to be disabled for now. When autoprefixer is
available, those lines can be removed and styleint with fix param should
be run to flush them in files.

In the meantime, stylelint-no-unsupported-browser-features plugin can be
used to highlight obsolete vendor prefixes.